requirements.txt
playwright>=1.50
requests>=2.30
python-dotenv>=1.0
2captcha-python>=2.0
Faker>=20.0
